Перенесла BLANK_TASK в edit-point, внесла hasOpenForm внутрь renderTask
