+++ src/view/edit-point.js
@@ -3,8 +3,45 @@ import OffersSelectorView from './offers-selector';
 import DestinationListView from './destination-list';
 import PhotosContainerView from './photos-container';
 import { createElement } from '../utils';
-import { BLANK_TASK } from '../consts';
 
+export const BLANK_TASK = {
+  basePrice: 0,
+  dateFrom: new Date(),
+  dateTo: new Date(),
+  destination: {
+    name: 'Geneva',
+    description:
+      'Geneva is a city in Switzerland that lies at the southern tip of expansive Lac Léman (Lake Geneva). Surrounded by the Alps and Jura mountains, the city has views of dramatic Mont Blanc.',
+    pictures: [
+      { src: 'img/photos/1.jpg', description: 'Event photo' },
+      { src: 'img/photos/2.jpg', description: 'Event photo' },
+      { src: 'img/photos/3.jpg', description: 'Event photo' },
+      { src: 'img/photos/4.jpg', description: 'Event photo' },
+      { src: 'img/photos/4.jpg', description: 'Event photo' },
+    ],
+  },
+  id: '',
+  type: 'flight',
+  offers: [
+    {
+      title: 'Add luggage',
+      price: 30,
+    },
+    {
+      title: 'Switch to comfort class',
+      price: 100,
+    },
+    {
+      title: 'Add meal',
+      price: 15,
+    },
+    {
+      title: 'Choose seats',
+      price: 5,
+    },
+  ],
+  isFavorite: false,
+};
 const createEditPointTemplate = (task) => {
   const { basePrice, dateFrom, destination, dateTo, type } = task;
   const iconPath = `img/icons/${type}.png`;
